+++ packages/react-magma-dom/src/components/Input/Input.stories.tsx
@@ -3,6 +3,8 @@ import { Input, InputProps } from '.';
 import { InputIconPosition, InputSize } from '../InputBase';
 import { Story, Meta } from '@storybook/react/types-6-0';
 import { NotificationsIcon } from 'react-magma-icons';
+import { Card, CardBody } from '../Card';
+import { magma } from '../../theme/magma';
 
 const Template: Story<InputProps> = args => (
   <Input {...args} labelText="Example" />
@@ -72,3 +74,18 @@ ClickableIconLarge.args = {
   ...ClickableIcon.args,
   inputSize: InputSize.large,
 };
+
+export const Inverse = Template.bind({});
+Inverse.args = {
+  ...Default.args,
+  isInverse: true,
+};
+Inverse.decorators = [
+  Story => (
+    <Card background={magma.colors.foundation} isInverse>
+      <CardBody>
+        <Story />
+      </CardBody>
+    </Card>
+  ),
+];
