+++ packages/react-magma-dom/src/index.ts
@@ -221,3 +221,5 @@ export { InverseContext, useIsInverse } from './inverse';
 export * from './components/Datagrid';
 export * from './components/Transition';
 export * from './components/Drawer';
+
+export * from './components/FormField';
