+++ website/react-magma-docs/src/pages/api/form-field-wrapper.mdx
@@ -1,0 +1,66 @@
+---
+pageTitle: FormField API
+title: FormField
+props:
+  - FormFieldProps
+---
+
+<DocsHeading type="api">FormField</DocsHeading>
+
+Standard container for form field, containing a label and helper or error message
+
+## Basic Usage
+
+```typescript
+import React from 'react';
+import { FormField } from 'react-magma-dom';
+
+export function Example() {
+  return <FormField>FormField content</FormField>;
+}
+```
+
+## InverseContext
+
+The `isInverse` prop will render a dark background and light text. The children will
+inherit the `isInverse` prop unless specified otherwise.
+
+```typescript
+import React from 'react';
+import { FormField } from 'react-magma-dom';
+
+export function Example() {
+  return <FormField isInverse>FormField content</FormField>;
+}
+```
+
+## Internationalization
+
+If you're using strings in your component, they must be internationalized.
+
+Full example of <Link to="/api/internationalization">internationalization override options</Link>
+
+```tsx
+import React from 'react';
+import { es } from 'date-fns/locale';
+import { DatePicker, I18nContext, defaultI18n } from 'react-magma-dom';
+
+export function Example() {
+  return (
+    <I18nContext.Provider
+      value={{
+        ...defaultI18n,
+        example: 'example i18n',
+      }}
+    >
+      <FormField>override default i18n value: </FormField>
+    </I18nContext.Provider>
+  );
+}
+```
+
+# FormField Props
+
+**Any other props supplied will be provided to the wrapping `div` element.**
+
+<FormFieldProps />
